feat: implement WebSocket server with client management and broadcasting
- Add ClientManager for handling subscriptions and broadcasting messages to clients.
- Introduce WsBusSink for routing serialized record updates to the ClientManager.
- Create WsDispatch for managing WebSocket sessions and handling authentication.
- Implement HTTP server with WebSocket upgrade and health check endpoints.
- Add StreamableRegistry for managing schema names of streamable types.
- Define reusable session handler traits for query handling and snapshot provision.
- Implement tests for ClientManager and StreamableRegistry functionalities.
