+++ aimdb-websocket-connector/src/server/auth.rs
